initial project structure setup
the project is split into two main parts: a library implementing the
logic and applications which use the library, either feeding in live
images from a camera or static images (for manual testing).
this way the library can also be covered with automated test coverage.

this is roughly based on the recommended structure in [Modern CMake].

the two apps currently print to `stderr` and return `1` instead of `0`
to indicate that they're not yet implemented. the library itself
currently does not contain any APIs.
the APIs and the implementation will follow later.

[Modern CMake]: https://cliutils.gitlab.io/modern-cmake/chapters/basics/structure.html