add test infrastructure based on GoogleTest
this uses [GoogleTest] (aka GTest) and integrates it using [CTest] into
CMake. the GTest sources are included using [vcpkg] - make sure to set
it up in your IDE so that everything works out of the box.
note that CMake comes with built-in tooling to include GTest tests in
CTest, see the [GoogleTest CMake documentation] and the [GTest CMake
quickstart].

as there's no API yet to be tested this is just an example - the name of
the test file & executable will have to be renamed once the first API is
being added!

[GoogleTest]: https://google.github.io/googletest
[CTest]: https://cmake.org/cmake/help/book/mastering-cmake/chapter/Testing%20With%20CMake%20and%20CTest.html
[vcpkg]: https://vcpkg.io/
[GoogleTest CMake documentation]: https://cmake.org/cmake/help/latest/module/GoogleTest.html
[GTest CMake quickstart]: https://google.github.io/googletest/quickstart-cmake.html