set up GitHub Actions for CI
this adds a GH Action so that any PRs & pushes get tested. it builds
everything on Windows & Linux using MSVC (Windows only) and GCC & CLang
(Linux only) and runs the tests.

this is losely based on the example action provided by GitHub for CMake
projects and also uses [run-vcpkg].

[dependabot] has been enabled for the GH actions as well to ensure that
they stay up to date. note that it can't be enabled for vcpkg as it
doesn't support it yet.

[run-vcpkg]: https://github.com/marketplace/actions/run-vcpkg
[dependabot]: https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/configuring-dependabot-version-updates