install necessary dependencies in CI
as we're using C++23 features we need to have newer versions of GCC &
CLang installed. furthermore, some of the dependencies of our
dependencies need certain packages to be installed in order to build
successfully.

this has been identified using trial & error.

note that the current `ubuntu-latest` is based on Ubuntu 22.04 LTS, the
full list of installed software can be found on the GitHub
documentation: [list of installed software].

[list of installed software]: https://github.com/actions/runner-images/blob/main/images/ubuntu/Ubuntu2204-Readme.md