CI: remove clang from build matrix
somehow it pulls in an older libc++ version which doesn't support C++23
yet (even though all required features are marked as supported in the
installed version). since we don't _need_ to have cross-platform and
cross-compiler support for this project it's ok to just drop clang
support for now (at least in the CI).

once Ubuntu 24.04 is released (planned for late April) and GitHub
supports it in the runner (unknown date) it might be possible to add
back support as it should ship with the new gcc & clang versions out of
the box.