add test images
these have been taken with our target camera and in the target setup,
i.e. camera mounted on a stand with 50cm between the lens and the ground
and a ring light around the lens.

this includes pictures of bananas as they're expected and additionally
paratial pictures as well as an empty picture and one with a meter on it
(to have a reference for the size).