`AnalysisError`: add convenience method `ToString`
this can be used when logging out the error to provide more context.
sadly it is not possible to have member methods on `enum`s in C++
(unlike other languages like Java, Rust, etc.), thus a workaround is
needed: an outer `class` provides the name and an inner `enum` provides
the values (as it is not an `enum class` the constants are visible as
members of the class).

this is possible due to the different overloaded operators to allow the
outer class to act transparently just like a real enum.

this is based on [this stackoverflow answer][so].

[so]: https://stackoverflow.com/a/53284026