implement `static-image` app
this adds a first version of the implementation which just reads a
single image (path provided as an argument), processes that with the
library and shows the annotated image at the end.

this is very unsophisticaed and it does not yet show any analysis result
(mainly because there are no results yet, `AnalysisResult` currently has
no content).

in a first step we can then just start printing the analysis result to
the console (to provide additional information over what's already
visible in the annotated image).
a (much) fancier version of the application could provide its own UI
with additional controls and embed the OpenCV image in there rather than
letting OpenCV handle the window controls.

note that we do not need to explicitly link OpenCV again in our app as
we already link `banana-lib` which defines the OpenCV dependency as a
public link, thus it's propagated to any target which links to it). the
same is also the case for the include directories.

this has been tested by temporarily not returning an error from
`AnalyzeImage` (not commited as part of this commit). the application
will work once the first step of the implementation has been done.

closes #4