implement `livecam` app
this is very similar to the `static-image` app.

note that the name is not 100% correct, as it does not just allow using
a live camera but it also allows using a video feed instead, as
supported by OpenCVs `VideoCapture` API which offers both options (when
passing an integer it is the index of the capture device, when passing a
string it's a file path, URL or similar to be opened).

just as with the `static-image` commit this has also been tested by
temporarily not failing in the `AnalyzeImage` method and it also does
not yet print any further results to the console (though printing to the
console when processing a lot of frames per second is suboptimal).

closes #5