wrap API methods in class
in the future (for #7) we'll have to know certain intrinsic and/or
extrinsic parameters of the camera (e.g. the distance between two pixels
on the image). this kind of information is best stored as private
members of a class and either provided upon initialization or as part of
a calibration routine.

while this commit doesn't yet add that information or any usage of it, it
already changes the API to be a class which has to be instantiated.