add internal APIs for contour detection & annotate image
this adds the necessary APIs to implement the contour detection of the
bananas and then use this information in a next step to annotate the
image.

further processing will be added in follow-up commits (to implement the
actual contour detection as well as further analysis of the detected
bananas).

a unit test for the image annotation cannot be added yet as the contour
detection is not yet implemented and returns only an empty list of
contours.