identify banana ripeness
this gives a rough estimation of the ripeness of the banana, with 100%
being considered "ripe", anything below it unripe and anything above it
over-ripe.
a very simple heuristic is used for this: if the banana is 100% yellow
it is ripe, any green colour part is considered unripe and any brown
colour part is considered over-ripe. thus the formula is simply:
$1 - g + b$, with $g$ and $b$ being the green and brown parts of the
whole banana.

solves #8