add library for 2D polyfitting
this uses the [ceres solver] to implement polyfitting using least
squares for a given set of 2D points to find a two-dimensional
polynomial ($y = a0 + a1 * x + a2 * x^2$).

this can then be used in a next step to fit the line through the banana
contour (will be done in a follow-up commit).

note that the `polyfit` header-only library is not specific to the
current project and could easily be factored out (it has been built in a
way to not rely on the rest of the project). it could potentially also
be made more abstract to support generic polyfitting for n-dimensional
polynomials and/or n-dimensional data points.

the API accepts a `range` to accept any form of iterable container. this
requires at least C++20, most likely C++23 (where ranges have been
updated significantly). it has only been tested against C++23. see the
[`std::ranges` documentation] for further details.

note that this is using ceres solver v2.1.0 while v2.2.0 has already
been released. the reason for this is that v2.1.0 is the latest
available version in vcpkg, see microsoft/vcpkg#34483.

to better show what is happening a jupyter notebook is provided,
visualising the solution approach using python. due to this `.gitignore`
had to be updated as jupyterlab stores additional cache files.

[ceres solver]: http://ceres-solver.org/
[`std::ranges` documentation]: https://en.cppreference.com/w/cpp/ranges