apps: report number of bananas found in console
this helps with seeing at a glance whether there were additional
artefacts being incorrectly identified as bananas (=> good reason to
change `Analyzer.match_max_score_` in `lib.hpp`).
once we have further information in the analysis result we can then also
print that out.

note that for videos we cannot print this on every frame as it'd
completely spam the log - instead the idea is to only print it when the
user presses 'i' (for 'information').

minor cosmetic things were smuggled into this commit as well (the import
order in the apps has been reordered to be alphabetical).