reduce threshold for shape matching to 0.6
with `resources/test-images/banana-00.jpg` it so far incorrectly
detected two bananas even though there's only one.

partial test coverage for this has been added, the rest of it (matching
against a reference version of the annotated image) will be done
separately as this is still an outstanding point from the original PR
which introduced the banana detection algorithm.