calculate coefficients for banana center line
this makes use of the polyfit library introduced in commit a6d96bd.
this also introduces the concept of verbose annotations: if enabled, the
center line is also drawn. more such information might follow in the
future. for now this is enabled for both apps.

note that this currently only works for bananas which are more-or-less
horizontally aligned - bananas which are vertically aligned incorrectly
lead to a line passing through the short side of it. this will have to
be addressed in a follow-up. the `banana-22.jpg` test image can be used
to reproduce this.