CI: skip testing on ubuntu until 24.04 is avialable on GHA
the new code now requires `std::views` which is not yet supported in GCC
13.1 (available on Ubuntu 22.04, the current `ubuntu-latest` in GHA). as
it's quite cumbersome to install GCC 13.2 (which adds support for this
and has been released nearly a year ago) the easiest option for now to
move forward is to just disable the Ubuntu tests. GHA will soon have the
new Ubuntu 24.04 available for beta-testing
(see actions/runner-images#9691) at which point we can try to re-enable
it based on the new beta. this should then also enable us to start
testing with CLang again.

for the time being tests will only run on Windows against MSVC (which is
new enough and supports all needed C++23 features).