refactor printing of analysis result into `<<` operator
this provides an `<<` operator overload for `ostream` for
`AnnotatedAnalysisResult` so that instead of the two apps manually
implementing the whole printing (with the risk of divergence between the
two) you can now simply write `std::cout << result;`.

this also fixes the fact that in the `livecam` app the rotation angle
hadn't been printed as the two apps diverged in the implementation.