handle rotated bananas
so far the center line had only been calculated correctly if the banana
was lying more-or-less horizontally.
however, thanks to the PCA analysis we know the rotation angle and
center of the banana. we can use this to rotate the contour first so
that the banana is considered horizontal (length-wise), which allows us
to correctly fit a polynomial through it to approximate the center line.

for rendering we then have to transform back the line to put it over the
banana again.