calculate mean curvature of banana
this calculates the mean geometric curvature of the center line of the
banana.

see [the wikipedia article] for the formula (or check the math 1 course
works). note that we ignore the direction of the curve as it is
irrelevant, i.e. we always return a positive curvature.

note also that the curvature is currently given in pixels as we do not
(yet) have any information about the physical size of objects on the
image (would need extrinsic camera calibration).

[the wikipedia article]: https://en.wikipedia.org/wiki/Curvature#Graph_of_a_function