add pixel to meter mapping and give measurements in meter
this now gives real numbers rather than pixels which are hard to judge
and depend on the pictures.

note that we currently do not do any automated extrinsic calibration,
accordingly the mapping is just hardcoded. different camera settings,
camera positions and of course camera models will require different
settings.

in the perfect world we'd have a setup routine which would scan for a
reference measurement (e.g. QR code) on the picture and automatically
calculate the size from this (either once on startup or on each
picture), but for this project this is overkill.

just remember to adjust the setting if you're using a picture taken with
a different camera - if you don't the numbers won't mean much :)

this is part of #7