CI: build with GCC 14 on Ubuntu 24.04
GHA now supports Ubuntu 24.04 as a beta (will become `ubuntu-latest`
once 24.04 LTS is released in autumn). with GCC 14 the project compiles
fine. it _should_ also compile with CLang 18, however i didn't get that
to work as Ubuntu ships with `libstdc++` (the C++ standard library from
GCC) which requires a newer version of "concepts" for `std::expected`
to work than what CLang implements - it should work with `libc++` (the
C++ standard library from CLang/LLVM), but that'd be much harder to set
up and isn't worth it now.